refactor: Remove commented code and update cron job frequency in task scheduler

+++ controllers/wrikeWebworkController.js
@@ -9,7 +9,6 @@ exports.handleWebhook = async (req, res) => {
     const taskID = req.body[0].taskId;
     console.log("Task ID:", taskID);
 
-
     if (req.body[0].eventType === "TaskStatusChanged") {
       // Get task details from Wrike
       const taskDetails = await wrikeService.getTaskDetails(taskID);
@@ -83,37 +82,44 @@ exports.handleWebhook = async (req, res) => {
         //get webwork task id from db by task.id
         const webworkTask = await webworkService.getWebworkTaskId(task.id);
         console.log("Webwork Task ID:", webworkTask);
+        if (webworkTask) {
+          const webworkUserId = webworkTask.webworkUserId;
+          //get webwork projectid from db by task.projectId
+          const webworkProjectId = await webworkService.getWebworkProjectID(
+            task.projectID
+          );
+          console.log("Webwork Project ID:", webworkProjectId);
 
-        const webworkUserId = webworkTask.webworkUserId;
-        //get webwork projectid from db by task.projectId
-        const webworkProjectId = await webworkService.getWebworkProjectID(
-          task.projectID
-        );
-        console.log("Webwork Project ID:", webworkProjectId);
+          // delete the task from db by task.id
+          await webworkService.deleteWebworkTask(webworkTask.webworkTaskId);
 
-        // delete the task from db by task.id
-        await webworkService.deleteWebworkTask(webworkTask.webworkTaskId);
+          //delete the task from webwork by task.id
+          await webworkService.deleteWebworkTaskFromWebwork(
+            webworkTask.webworkTaskId
+          );
 
-        //delete the task from webwork by task.id
-        await webworkService.deleteWebworkTaskFromWebwork(
-          webworkTask.webworkTaskId
-        );
+          // get all tasks from task db by userId
+          await webworkService.removeAssignee(webworkUserId, webworkProjectId);
 
-        // get all tasks from task db by userId
-        await webworkService.removeAssignee(webworkUserId, webworkProjectId);
+          setTimeout(async () => {
+            const timeSpentinwebwork = await webworkService.getTaskTime(
+              webworkTask
+            );
+            console.log("Time Spent:", timeSpent);
 
-        setTimeout(async () => {
-          const timeSpent = await webworkService.getTaskTime(webworkTask);
-          console.log("Time Spent:", timeSpent);
+            const timeSpentfromdb = webworkTask.timeSpent;
 
-          // todays date yyyy-mm-dd
-          const date = new Date().toISOString().split("T")[0];
-          console.log("Date:", date);
+            const timeSpent = timeSpentinwebwork - timeSpentfromdb;
 
-          //call this api /tasks/{task.id}/timelogs?hours={timespent/3600}&trackedDate={dateOnly}
-          if (timeSpent > 0)
-            await wrikeService.updateTaskTime(task.id, timeSpent, date);
-        }, 10 * 60 * 1000); // wait for 10 minute before getting time of task in webwork
+            // todays date yyyy-mm-dd
+            const date = new Date().toISOString().split("T")[0];
+            console.log("Date:", date);
+
+            //call this api /tasks/{task.id}/timelogs?hours={timespent/3600}&trackedDate={dateOnly}
+            if (timeSpent > 0)
+              await wrikeService.updateTaskTime(task.id, timeSpent, date);
+          }, 10 * 60 * 1000); // wait for 10 minute before getting time of task in webwork
+        }
       }
     }
